refactor(core): move genAggregateEpochKey proof to synchronizer
